requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Async support
aiofiles==23.2.1
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Printer communication
pycups==2.0.1

# System utilities
psutil==5.9.8
netifaces==0.11.0

# Logging and monitoring
python-json-logger==2.0.7
